# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comparador de Normativas — API REST & SSE",
...
    "version": "4.0.0"
  },
  "paths": {
...
    "/api/documents/{tipo}/{doc_id}/pdf": {
      "get": {
        "tags": [
          "Documentos"
        ],
        "summary": "Get Document Pdf",
        "description": "Entrega el archivo PDF binario para renderizado en PDF.js con soporte de byte-ranges.",
        "operationId": "get_document_pdf_api_documents__tipo___doc_id__pdf_get",
        "parameters": [
          {
            "name": "tipo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tipo"
            }
          },
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "tags": [
          "Documentos"
        ],
        "summary": "Upload Document",
        "description": "Sube un documento PDF al almacenamiento de uploads.",
        "operationId": "upload_document_api_documents_upload_post",
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tipo"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/tabular": {
      "post": {
        "tags": [
          "Documentos"
        ],
        "summary": "Tabulate Documents",
        "description": "Ejecuta la tabulación de documentos mediante Docling.",
        "operationId": "tabulate_documents_api_documents_tabular_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TabulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TabulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/compare/start": {
      "post": {
        "tags": [
          "Comparación"
        ],
        "summary": "Start Compare",
        "description": "Inicia una corrida de comparación asíncrona.",
        "operationId": "start_compare_api_compare_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartCompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare/stream/{run_id}": {
      "get": {
        "tags": [
          "Comparación"
        ],
        "summary": "Stream Progress",
        "description": "Canal Server-Sent Events (SSE) para recibir el progreso en vivo.",
        "operationId": "stream_progress_api_compare_stream__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare/runs/{run_id}": {
      "get": {
        "tags": [
          "Comparación"
        ],
        "summary": "Get Run Status",
        "description": "Retorna el estado detallado y resultados de una corrida.",
        "operationId": "get_run_status_api_compare_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare/runs/{run_id}/export/{formato}": {
      "get": {
        "tags": [
          "Comparación"
        ],
        "summary": "Export Run",
        "description": "Descarga el Papel de Trabajo en Excel (6 hojas) o archivo JSON.",
        "operationId": "export_run_api_compare_runs__run_id__export__formato__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "formato",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Formato"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunScopeModel": {
        "properties": {
          "articulos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Articulos"
          },
          "secciones": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secciones"
          },
          "doc_ids_normativa": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Doc Ids Normativa"
          },
          "preset_articulos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Articulos"
          },
          "muestra_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Muestra N"
          },
          "muestra_aleatoria": {
            "type": "boolean",
            "title": "Muestra Aleatoria",
            "default": false
          }
        },
        "type": "object",
        "title": "RunScopeModel"
      },
      "RunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "progreso": {
            "type": "integer",
            "title": "Progreso"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "porcentaje": {
            "type": "number",
            "title": "Porcentaje"
          },
          "etiqueta": {
            "type": "string",
            "title": "Etiqueta",
            "default": ""
          },
          "cobertura_global": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cobertura Global"
          },
          "alerta_cobertura": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alerta Cobertura"
          },
          "articulos_sin_cobertura": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Articulos Sin Cobertura"
          },
          "vista_manual": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Vista Manual"
          },
          "vista_normativa": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Vista Normativa"
          },
          "motivos_revision": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Motivos Revision"
          },
          "total_revision_manual": {
            "type": "integer",
            "title": "Total Revision Manual",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "estado",
          "progreso",
          "total",
          "porcentaje"
        ],
        "title": "RunStatusResponse"
      },
...
      "StartCompareRequest": {
        "properties": {
          "dual_mode": {
            "type": "boolean",
            "title": "Dual Mode",
            "default": true
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunScopeModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "min_score": {
            "type": "number",
            "title": "Min Score",
            "default": 0.3
          },
          "top_k": {
            "type": "integer",
            "title": "Top K",
            "default": 5
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "embed_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Model"
          }
        },
        "type": "object",
        "title": "StartCompareRequest"
      },
      "StartCompareResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "stream_url"
        ],
        "title": "StartCompareResponse"
      },
      "TabulateRequest": {
        "properties": {
          "normativas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Normativas"
          },
          "manuales": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Manuales"
          },
          "do_ocr": {
            "type": "boolean",
            "title": "Do Ocr",
            "default": false
          }
        },
        "type": "object",
        "title": "TabulateRequest"
      },
      "TabulateResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "normativa_elementos": {
            "type": "integer",
            "title": "Normativa Elementos"
          },
          "manual_secciones": {
            "type": "integer",
            "title": "Manual Secciones"
          },
          "elapsed_seconds": {
            "type": "number",
            "title": "Elapsed Seconds"
          }
        },
        "type": "object",
        "required": [
          "status",
          "normativa_elementos",
          "manual_secciones",
          "elapsed_seconds"
        ],
        "title": "TabulateResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}